+++ src/include/file-manager.h
@@ -22,16 +22,16 @@ public:
 	 * (e.g. the file does not exist)
 	 * and true otherwise. 
 	 * */
-	static GraphRepresentation<unsigned int>* readFile(std::string filepath,
+	static GraphRepresentation<size_t>* readFile(std::string filepath,
 			FileType type, char separator=stdSeparator, char comment=stdComment);
 
 private:
 	static const char stdComment = '#';
 	static const char stdSeparator = ',';
 
-	static AdjacencyList<unsigned int>* readAdjList(std::ifstream* file,
+	static AdjacencyList<size_t>* readAdjList(std::ifstream* file,
 			char separator, char comment);
-	static AdjacencyMatrix<unsigned int>* readAdjMatrix(std::ifstream* file,
+	static AdjacencyMatrix<size_t>* readAdjMatrix(std::ifstream* file,
 			char separator, char comment);
 	
 	/* Parses a csv line.
@@ -45,7 +45,7 @@ private:
 	 *
 	 * The function returns the length of `result`.
 	 */
-	static unsigned int parseLine(std::string* line, int* values,
+	static size_t parseLine(std::string* line, int* values,
 			char separator, char comment);
 };
